requirements.txt
# Core scRNA-Seq analysis packages
scanpy>=1.10.0
anndata>=0.10.0

# Data manipulation and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter environment
jupyterlab>=4.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Additional utilities
scikit-learn>=1.3.0
umap-learn>=0.5.4
python-igraph>=0.11.0
leidenalg>=0.10.0
pyyaml>=6.0

# Optional: For enhanced visualizations
plotly>=5.17.0

# Optional: For parallel processing
joblib>=1.3.0

# Optional: For statistical testing
statsmodels>=0.14.0
